+++ src/features/edit/hooks/index.ts
@@ -1,5 +1,6 @@
 export { useProfileSchema } from './useProfileSchema';
 export { useRecordStatus } from './useRecordStatus';
 export { useResetRecordStatus } from './useResetRecordStatus';
+export { useResourceExport } from './useResourceExport';
 export { useSaveRecord } from './useSaveRecord';
 export { useSaveRecordWarning } from './useSaveRecordWarning';
