+++ src/features/resources/hooks/index.ts
@@ -1,2 +1,3 @@
 export { useRecordGeneration } from './useRecordGeneration';
 export { useProcessedRecordAndSchema } from './useProcessedRecordAndSchema';
+export { useMarcData } from './useMarcData';
